Refactor disk loop resource management methods
Consolidated shared methods like provisioner, unprovisioner, and _status into a common base implementation in `__init__.py`.
Removed redundant and platform-specific implementations from `sunos.py`, `freebsd.py`, `darwin.py`, and `linux.py` to
eliminate code duplication and simplify maintenance.
