+++ opensvc/drivers/resource/disk/loop/darwin.py
@@ -1,12 +1,10 @@
 import os
 
-import core.status
 import core.exceptions as ex
+import core.status
 import utilities.devices.darwin
 
 from . import BaseDiskLoop
-from utilities.converters import convert_size
-from core.objects.svcdict import KEYS
 
 DRIVER_GROUP = "disk"
 DRIVER_BASENAME = "loop"
@@ -50,42 +48,11 @@ class DiskLoop(BaseDiskLoop):
                 raise ex.Error
 
     def _status(self, verbose=False):
+        r = self.svc.resource_handling_file(self.loopfile)
+        if self.is_provisioned() and not os.path.exists(self.loopfile):
+            if r is None or (r and r.status() in (core.status.UP, core.status.STDBY_UP)):
+                self.status_log("%s does not exist" % self.loopfile)
         if self.is_up():
             return core.status.UP
         else:
             return core.status.DOWN
-
-    def provisioned(self):
-        try:
-            return os.path.exists(self.loopfile)
-        except Exception:
-            return
-
-    def unprovisioner(self):
-        try:
-            self.loopfile
-        except Exception as e:
-            raise ex.Error(str(e))
-
-        if not self.provisioned():
-            return
-
-        self.log.info("unlink %s" % self.loopfile)
-        os.unlink(self.loopfile)
-        self.svc.node.unset_lazy("devtree")
-
-    def provisioner(self):
-        d = os.path.dirname(self.loopfile)
-        try:
-            if not os.path.exists(d):
-                self.log.info("create directory %s"%d)
-                os.makedirs(d)
-            with open(self.loopfile, 'w') as f:
-                self.log.info("create file %s, size %s"%(self.loopfile, self.size))
-                f.seek(convert_size(self.size, _to='b', _round=512)-1)
-                f.write('\0')
-            self.chown()
-            self.chmod()
-        except Exception as e:
-            raise ex.Error("failed to create %s: %s"% (self.loopfile, str(e)))
-        self.svc.node.unset_lazy("devtree")
