+++ opensvc/drivers/resource/disk/loop/linux.py
@@ -8,7 +8,6 @@ import utilities.devices.linux
 from . import BaseDiskLoop
 from env import Env
 from utilities.cache import clear_cache
-from utilities.converters import convert_size
 from utilities.lock import cmlock
 from utilities.mounts.linux import Mounts
 from utilities.files import getmount
@@ -92,12 +91,8 @@ class DiskLoop(BaseDiskLoop):
         if os.path.exists(self.loopfile):
             self.auto_unprovision()
 
-    def resource_handling_file(self):
-        path = os.path.dirname(self.loopfile)
-        return self.svc.resource_handling_dir(path)
-
     def _status(self, verbose=False):
-        r = self.resource_handling_file()
+        r = self.svc.resource_handling_file(self.loopfile)
         if self.is_provisioned() and not os.path.exists(self.loopfile):
             if r is None or (r and r.status() in (core.status.UP, core.status.STDBY_UP)):
                 self.status_log("%s does not exist" % self.loopfile)
@@ -109,38 +104,3 @@ class DiskLoop(BaseDiskLoop):
     def exposed_devs(self):
         self.loop = utilities.devices.linux.file_to_loop(self.loopfile)
         return set(self.loop)
-
-    def provisioned(self):
-        try:
-            return os.path.exists(self.loopfile)
-        except Exception:
-            return
-
-    def unprovisioner(self):
-        try:
-            self.loopfile
-        except Exception as e:
-            raise ex.Error(str(e))
-
-        if not self.provisioned():
-            return
-
-        self.log.info("unlink %s" % self.loopfile)
-        os.unlink(self.loopfile)
-        self.svc.node.unset_lazy("devtree")
-
-    def provisioner(self):
-        d = os.path.dirname(self.loopfile)
-        try:
-            if not os.path.exists(d):
-                self.log.info("create directory %s"%d)
-                os.makedirs(d)
-            with open(self.loopfile, 'w') as f:
-                self.log.info("create file %s, size %s"%(self.loopfile, self.size))
-                f.seek(convert_size(self.size, _to='b', _round=512)-1)
-                f.write('\0')
-            self.chown()
-            self.chmod()
-        except Exception as e:
-            raise ex.Error("failed to create %s: %s"% (self.loopfile, str(e)))
-        self.svc.node.unset_lazy("devtree")
