+++ opensvc/drivers/resource/disk/loop/sunos.py
@@ -6,10 +6,8 @@ import core.status
 import utilities.devices.sunos
 
 from . import BaseDiskLoop
-from utilities.converters import convert_size
 from utilities.lock import cmlock
 from env import Env
-from core.objects.svcdict import KEYS
 
 DRIVER_GROUP = "disk"
 DRIVER_BASENAME = "loop"
@@ -60,12 +58,8 @@ class DiskLoop(BaseDiskLoop):
             if ret != 0:
                 raise ex.Error
 
-    def resource_handling_file(self):
-        path = os.path.dirname(self.loopfile)
-        return self.svc.resource_handling_dir(path)
-
     def _status(self, verbose=False):
-        r = self.resource_handling_file()
+        r = self.svc.resource_handling_file(self.loopfile)
         if self.is_provisioned() and not os.path.exists(self.loopfile):
             if r is None or (r and r.status() in (core.status.UP, core.status.STDBY_UP)):
                 self.status_log("%s does not exist" % self.loopfile)
@@ -77,38 +71,3 @@ class DiskLoop(BaseDiskLoop):
     def exposed_devs(self):
         self.loop = utilities.devices.sunos.file_to_loop(self.loopfile)
         return set(self.loop)
-
-    def provisioned(self):
-        try:
-            return os.path.exists(self.loopfile)
-        except Exception:
-            return
-
-    def unprovisioner(self):
-        try:
-            self.loopfile
-        except Exception as e:
-            raise ex.Error(str(e))
-
-        if not self.provisioned():
-            return
-
-        self.log.info("unlink %s" % self.loopfile)
-        os.unlink(self.loopfile)
-        self.svc.node.unset_lazy("devtree")
-
-    def provisioner(self):
-        d = os.path.dirname(self.loopfile)
-        try:
-            if not os.path.exists(d):
-                self.log.info("create directory %s"%d)
-                os.makedirs(d)
-            with open(self.loopfile, 'w') as f:
-                self.log.info("create file %s, size %s"%(self.loopfile, self.size))
-                f.seek(convert_size(self.size, _to='b', _round=512)-1)
-                f.write('\0')
-            self.chown()
-            self.chmod()
-        except Exception as e:
-            raise ex.Error("failed to create %s: %s"% (self.loopfile, str(e)))
-        self.svc.node.unset_lazy("devtree")
